Fix imports after moving plot functions
